Renamed footer links to social links
