--- src/components/icons/instagram-icon/instagram-icon.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { SvgIcon } from '@material-ui/core';
-import { useStyles } from './instagram-icon.styles';
-
-export const InstagramIcon = ({ color }) => {
-  const styles = useStyles();
-  return (
-    <SvgIcon className={styles.icon} viewBox='0 0 40 40' fill='none'>
-      <path
-        d='M20.0001 17.3963C18.5665 17.3963 17.3966 18.5663 17.3966 19.9998C17.3966 21.4334 18.5665 22.6034 20.0001 22.6034C21.4337 22.6034 22.6036 21.4334 22.6036 19.9998C22.6036 18.5663 21.4337 17.3963 20.0001 17.3963ZM27.8087 19.9998C27.8087 18.9217 27.8185 17.8534 27.7579 16.7772C27.6974 15.5272 27.4122 14.4178 26.4981 13.5038C25.5821 12.5877 24.4747 12.3045 23.2247 12.244C22.1466 12.1834 21.0782 12.1932 20.002 12.1932C18.9239 12.1932 17.8556 12.1834 16.7794 12.244C15.5294 12.3045 14.42 12.5897 13.506 13.5038C12.5899 14.4198 12.3067 15.5272 12.2462 16.7772C12.1856 17.8553 12.1954 18.9237 12.1954 19.9998C12.1954 21.076 12.1856 22.1463 12.2462 23.2225C12.3067 24.4725 12.5919 25.5819 13.506 26.4959C14.422 27.412 15.5294 27.6952 16.7794 27.7557C17.8575 27.8163 18.9259 27.8065 20.002 27.8065C21.0802 27.8065 22.1485 27.8163 23.2247 27.7557C24.4747 27.6952 25.5841 27.41 26.4981 26.4959C27.4142 25.5799 27.6974 24.4725 27.7579 23.2225C27.8204 22.1463 27.8087 21.078 27.8087 19.9998V19.9998ZM20.0001 24.0057C17.7833 24.0057 15.9942 22.2166 15.9942 19.9998C15.9942 17.7831 17.7833 15.994 20.0001 15.994C22.2169 15.994 24.006 17.7831 24.006 19.9998C24.006 22.2166 22.2169 24.0057 20.0001 24.0057ZM24.17 16.7655C23.6524 16.7655 23.2345 16.3475 23.2345 15.8299C23.2345 15.3123 23.6524 14.8944 24.17 14.8944C24.6876 14.8944 25.1056 15.3123 25.1056 15.8299C25.1057 15.9528 25.0816 16.0746 25.0347 16.1881C24.9877 16.3017 24.9188 16.4049 24.8319 16.4918C24.745 16.5787 24.6418 16.6476 24.5282 16.6946C24.4146 16.7415 24.2929 16.7656 24.17 16.7655V16.7655Z'
-        fill={color}
-      />
-      <circle opacity='0.3' cx='20' cy='20' r='19.6' stroke={color} strokeWidth='0.8' />
-    </SvgIcon>
-  );
-};
