--- src/containers/footer-links/index.js
@@ -1,3 +1,0 @@
-import FooterLinks from './footer-links';
-
-export default FooterLinks;
