+++ src/containers/social-links/index.js
@@ -1,0 +1,3 @@
+import SocialLinks from './social-links';
+
+export default SocialLinks;
